+++ packages/form-table/src/index.ts
@@ -1,0 +1,15 @@
+import type { PluginObject } from 'vue'
+import FormTable from './index.vue'
+
+export * from './types'
+export * from './configs'
+export { FormTable }
+
+const plugin: PluginObject<undefined> = {
+  install(Vue) {
+    Vue.component('FormTable', FormTable)
+  }
+}
+
+export default FormTable
+export { plugin as FormTablePlugin }