+++ packages/form-table/src/public-types.ts
@@ -1,0 +1,9 @@
+import type { Component, PluginObject } from 'vue'
+
+export * from './types'
+export * from './configs'
+
+export declare const FormTable: Component
+declare const _default: Component
+export default _default
+export declare const FormTablePlugin: PluginObject<undefined>