+++ playground/vite.config.ts
@@ -15,7 +15,8 @@ export default defineConfig({
   ],
   resolve: {
     alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url))
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+      formtable: fileURLToPath(new URL('../packages/form-table/src/index.ts', import.meta.url))
     }
   }
 })